requirements/base.txt
pytz==2022.7  # https://github.com/stub42/pytz
python-slugify==7.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.3.0  # https://github.com/evansd/whitenoise
redis==4.4.1  # https://github.com/redis/redis-py
hiredis==2.1.1  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.3  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-theme-pixel
django-bootstrap-v5
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.25.1  # https://github.com/tfranzel/drf-spectacular
